+++ src/global.d.ts
@@ -2,17 +2,8 @@
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
-declare const google: any
-
 interface ImportMetaEnv {
-  VITE_DATABASE_URL: string
   VITE_GOOGLE_CLIENT_ID: string
-  VITE_GOOGLE_SECRET: string
-  VITE_JWT_SECRET: string
-  VITE_WEB_URL: string
-  VITE_SEND_IN_BLUE_KEY: string
-  VITE_SEND_IN_BLUE_URL: string
-  VITE_EMAIL_FROM: string
 }
 
 type AuthenticationResult = {
@@ -55,4 +46,8 @@ type User = {
 type UserSession = {
   id: string,
   user: User
+}
+
+interface Window {
+  google?: any
 }